+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 - getLastActivity() method on BotContext
 - GroupBot class
+- GroupBotFactory class
 
 ## [0.46.5] - 2026-05-25